chore: share parquet readers and unify factory loading
