+++ src/decoding/current/logs.rs
@@ -10,7 +10,7 @@ use crate::decoding::logs::{
 };
 use crate::decoding::types::DecoderMessage;
 use crate::live::LiveStorage;
-use crate::raw_data::historical::eth_calls::read_factory_addresses_from_parquet;
+use crate::storage::parquet_readers::read_factory_addresses_from_parquet;
 use crate::transformations::{DecodedEventsMessage, RangeCompleteMessage};
 
 /// Load accumulated factory addresses from both compacted parquet and uncompacted bincode files.
