+++ src/raw_data/historical/catchup/eth_calls.rs
@@ -13,8 +13,7 @@ use crate::raw_data::historical::blocks::{
 use crate::raw_data::historical::eth_calls::{
     build_call_configs, build_event_triggered_call_configs, build_factory_once_call_configs,
     build_once_call_configs, build_token_call_configs, event_output_exists,
-    get_existing_log_ranges, load_factory_addresses_for_once_catchup,
-    load_historical_factory_addresses, load_or_build_once_column_index, process_event_triggers,
+    get_existing_log_ranges, load_or_build_once_column_index, process_event_triggers,
     process_event_triggers_multicall, process_factory_once_calls, process_once_calls_multicall,
     process_once_calls_regular, process_range, process_range_multicall, process_token_range,
     process_token_range_multicall, read_logs_from_parquet, read_once_column_index,
@@ -24,6 +23,9 @@ use crate::raw_data::historical::eth_calls::{
 use crate::raw_data::historical::factories::{get_factory_call_configs, FactoryAddressData};
 use crate::raw_data::historical::receipts::{build_event_trigger_matchers, extract_event_triggers};
 use crate::rpc::UnifiedRpcClient;
+use crate::storage::factory_data::{
+    load_factory_addresses_by_collection, load_factory_addresses_with_metadata,
+};
 use crate::storage::paths::raw_eth_calls_dir;
 use crate::storage::{DataLoader, S3Manifest, StorageManager};
 use crate::types::config::chain::ChainConfig;
@@ -526,9 +528,10 @@ pub async fn collect_eth_calls(
         }
 
         // Load factory address data from existing parquet files
-        let factory_catchup_data = load_factory_addresses_for_once_catchup(
+        let collections_needed: HashSet<String> = factory_once_configs.keys().cloned().collect();
+        let factory_catchup_data = load_factory_addresses_with_metadata(
             &chain.name,
-            &factory_once_configs,
+            &collections_needed,
             s3_manifest.as_ref(),
             storage_manager.as_ref(),
         )
@@ -618,9 +621,15 @@ pub async fn collect_eth_calls(
     if has_event_triggered_calls {
         // CRITICAL: Load historical factory addresses BEFORE processing event triggers
         // This ensures we can properly filter events from factory-created contracts
-        let historical_factory_addrs = load_historical_factory_addresses(
+        let factory_collections: HashSet<String> = event_call_configs
+            .values()
+            .flatten()
+            .filter(|c| c.is_factory)
+            .map(|c| c.contract_name.clone())
+            .collect();
+        let historical_factory_addrs = load_factory_addresses_by_collection(
             &chain.name,
-            &event_call_configs,
+            &factory_collections,
             s3_manifest.as_ref(),
             storage_manager.as_ref(),
         )
