+++ src/raw_data/historical/eth_calls/types.rs
@@ -27,6 +27,9 @@ pub enum EthCallCollectionError {
     #[error("Arrow error: {0}")]
     Arrow(#[from] arrow::error::ArrowError),
 
+    #[error("Parquet read error: {0}")]
+    ParquetRead(#[from] crate::storage::parquet_readers::ParquetReadError),
+
     #[error("Parameter error: {0}")]
     Param(#[from] ParamError),
